Changed variables with type double --> type float. Rewrote a few variable names for better readability

+++ game/Monster.java
@@ -8,7 +8,7 @@
  */
 public class Monster extends Entity {
     // Fields
-    private double hp;
+    private float hp;
     private String name;
     private Item weapon;
     protected EnemyType enemyType;
